+++ internal/services/smart_mobile_device_group_v1/resource_model.go
@@ -18,7 +18,7 @@ type smartMobileDeviceGroupV1ResourceModel struct {
 // smartMobileDeviceGroupV1CriteriaDataModel describes the criteria data model.
 type smartMobileDeviceGroupV1CriteriaDataModel struct {
 	Name         types.String `tfsdk:"name"`
-	Priority     types.Int32  `tfsdk:"priority"`
+	Priority     types.Int64  `tfsdk:"priority"`
 	AndOr        types.String `tfsdk:"and_or"`
 	SearchType   types.String `tfsdk:"search_type"`
 	Value        types.String `tfsdk:"value"`
