+++ internal/services/smart_mobile_device_group_v1/resource_state.go
@@ -33,7 +33,7 @@ func state(ctx context.Context, data *smartMobileDeviceGroupV1ResourceModel, res
 	for _, criterion := range resp.Criteria {
 		criteriaModel := smartMobileDeviceGroupV1CriteriaDataModel{
 			Name:       types.StringValue(criterion.Name),
-			Priority:   types.Int32Value(int32(criterion.Priority)),
+			Priority:   types.Int64Value(int64(criterion.Priority)),
 			AndOr:      types.StringValue(strings.ToLower(criterion.AndOr)),
 			SearchType: types.StringValue(criterion.SearchType),
 			Value:      types.StringValue(criterion.Value),
